+++ platform/api/azure/instance.go
@@ -23,7 +23,7 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2021-03-01/compute"
+	"github.com/Azure/azure-sdk-for-go/services/compute/mgmt/2022-08-01/compute"
 	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2021-02-01/network"
 
 	"github.com/flatcar/mantle/util"
@@ -123,7 +123,7 @@ func (a *API) getVMParameters(name, userdata, sshkey, storageAccountURI string, 
 						ID: nic.ID,
 						NetworkInterfaceReferenceProperties: &compute.NetworkInterfaceReferenceProperties{
 							Primary:      util.BoolToPtr(true),
-							DeleteOption: compute.DeleteOptionsDelete,
+							DeleteOption: compute.Delete,
 						},
 					},
 				},
@@ -135,6 +135,13 @@ func (a *API) getVMParameters(name, userdata, sshkey, storageAccountURI string, 
 				},
 			},
 		},
+	}
+
+	switch a.Opts.DiskController {
+	case "nvme":
+		vm.VirtualMachineProperties.StorageProfile.DiskControllerType = compute.NVMe
+	case "scsi":
+		vm.VirtualMachineProperties.StorageProfile.DiskControllerType = compute.SCSI
 	}
 
 	// I don't think it would be an issue to have empty user-data set but better